// apps/frontend/src/types/api.ts
export interface components {
    schemas: {
        SuccessResponse: {
            /** @example 0 */
            code?: number;
            /** @example ok */
            message?: string;
            data?: Record<string, never>;
        };
        ErrorResponse: {
            /** @example 40001 */
            code?: number;
            /** @example 错误描述 */
            message?: string;
            /** @example 详细错误信息 */
            detail?: string;
        };
// ...
        RefreshResponse: {
            accessToken?: string;
            expiresIn?: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}